refactor: move dailies from ecotracker/ to common/
