+++ app/src/main/java/com/b07/planetze/common/daily/food/MealDaily.java
@@ -1,11 +1,11 @@
-package com.b07.planetze.ecotracker.daily.food;
+package com.b07.planetze.common.daily.food;
 
 import androidx.annotation.NonNull;
 
 import com.b07.planetze.common.Emissions;
 import com.b07.planetze.util.measurement.Mass;
-import com.b07.planetze.ecotracker.daily.Daily;
-import com.b07.planetze.ecotracker.daily.DailyType;
+import com.b07.planetze.common.daily.Daily;
+import com.b07.planetze.common.daily.DailyType;
 
 public record MealDaily(
         @NonNull MealType mealType,
