+++ app/src/main/java/com/b07/planetze/common/daily/shopping/BuyElectronicsForm.java
@@ -1,8 +1,8 @@
-package com.b07.planetze.ecotracker.daily.shopping;
+package com.b07.planetze.common.daily.shopping;
 
 import androidx.annotation.NonNull;
 
-import com.b07.planetze.ecotracker.daily.DailyForm;
+import com.b07.planetze.common.daily.DailyForm;
 import com.b07.planetze.ecotracker.exception.DailyFormException;
 import com.b07.planetze.form.Form;
 import com.b07.planetze.form.FormSubmission;
