+++ app/src/main/java/com/b07/planetze/common/daily/transport/PublicTransitDaily.java
@@ -1,12 +1,12 @@
-package com.b07.planetze.ecotracker.daily.transport;
+package com.b07.planetze.common.daily.transport;
 
 import androidx.annotation.NonNull;
 
 import com.b07.planetze.common.Emissions;
 import com.b07.planetze.util.measurement.ImmutableDuration;
 import com.b07.planetze.util.measurement.Mass;
-import com.b07.planetze.ecotracker.daily.Daily;
-import com.b07.planetze.ecotracker.daily.DailyType;
+import com.b07.planetze.common.daily.Daily;
+import com.b07.planetze.common.daily.DailyType;
 
 public record PublicTransitDaily(
         @NonNull TransitType transitType,
