+++ app/src/main/java/com/b07/planetze/database/DatabaseError.java
@@ -5,9 +5,12 @@ import androidx.annotation.NonNull;
 import java.util.Objects;
 
 /**
- * An error that occurred while calling a {@link Database}.
+ * An error that occurred while calling a {@link Database} function.
  */
 public interface DatabaseError {
+    /**
+     * {@return the message associated with this error}
+     */
     @NonNull
     String message();
 }
