+++ app/src/main/java/com/b07/planetze/ecotracker/exception/DailyFormException.java
@@ -1,7 +1,9 @@
 package com.b07.planetze.ecotracker.exception;
 
+import com.b07.planetze.common.daily.DailyForm;
+
 /**
- * Thrown when a {@link com.b07.planetze.ecotracker.daily.DailyForm} failed to
+ * Thrown when a {@link DailyForm} failed to
  * create a {@link com.b07.planetze.form.Form}. <br>
  * This indicates an error in the form's definition.
  */
